requirements.txt
django==1.7.3

djangorestframework==3.0.3
djangorestframework-jwt==1.2.0

django-cors-headers==1.0.0
django-extensions==1.5.0
django-filter==0.9.2
django-permission==0.8.3
djangorestframework-composed-permissions==0.1

mysql-python==1.2.5
sqlparse==0.1.14
markdown==2.5.2
werkzeug==0.10.1
django-debug-toolbar==1.2.2
mock==1.0.1
